refactor: extract sub-function for batch
