+++ src/frontend/src/eth/services/batch-eth-transactions.services.ts
@@ -1,0 +1,38 @@
+import { ETHERSCAN_MAX_CALLS_PER_SECOND } from '$env/rest/etherscan.env';
+import { loadEthereumTransactions } from '$eth/services/eth-transactions.services';
+import { batch } from '$lib/services/batch.services';
+import type { Token, TokenId } from '$lib/types/token';
+import type { ResultSuccess } from '$lib/types/utils';
+
+export type ResultByToken = ResultSuccess & { tokenId: TokenId };
+
+type PromiseResult = Promise<ResultByToken>;
+
+export const batchLoadTransactions = ({
+	tokens,
+	tokensAlreadyLoaded
+}: {
+	tokens: Token[];
+	tokensAlreadyLoaded: TokenId[];
+}): AsyncGenerator<PromiseSettledResult<ResultByToken>[]> => {
+	const promises = tokens.reduce<PromiseResult[]>(
+		(acc, { network: { id: networkId }, id: tokenId }) => {
+			if (tokensAlreadyLoaded.includes(tokenId)) {
+				return acc;
+			}
+
+			const promise = (async (): PromiseResult => {
+				const result = await loadEthereumTransactions({ tokenId, networkId });
+				return { ...result, tokenId };
+			})();
+
+			return [...acc, promise];
+		},
+		[]
+	);
+
+	return batch<ResultByToken>({
+		promises,
+		batchSize: ETHERSCAN_MAX_CALLS_PER_SECOND
+	});
+};
